auto-create topics (#202)
Topics are now allowed be auto-created when using the Quix Platform (via Application.Quix()) when calling .run().

Since auto-creating topics is not a valid option for non-Quix apps, the variable is stored by the QuixKafkaConfigsBuilder, and is set by the Application.Quix(auto_create_topics=BOOL) (default=True).

If auto create is True, create any missing topics, or validate any that are pending (status=="Creating)". There is a brief period we must "validate" them, aka wait for them to have a status=="Ready".
If auto create is False, confirm that the expected topics exist, else raise an exception.

In addition, Application.topic now has a new `creation_configs` arg (just for Quix apps), which accepts a new `TopicCreationConfigs` object (from `platforms/quix/config.py`). This can be used to pass configs for a creating a topic; one gets generated for you by default with config defaults that match the platform defaults. These objects are also stored on `QuixKafkaConfigsBuilder`.