add some examples (uses auto-create-topics) (#203)
Created some examples along with some boilerplate/helpers to easily get started. This one includes a `bank_example` with a few different serialization/platform approaches. The example has a `README.md` in its respective folder for more details.

There are some .env files for users to fill out as needed. Unfortunately there's no good way to force files to be otherwise untracked while also including them in the repo (at least not without the user doing something themselves). Not thrilled with the solution but it's probably fine for now.

See StreamingDataFrame/examples/README.md for general details on how to setup and run the examples.