Stateful sources implementation (#615)
Enhance source connectors by allowing access to a kafka backed state

* Implement stateful sources

* Added support for creating changelog topics without a related "data" topic

Stateful sources don't have an input topic on which to associate the changelog topic. The topic and recovery manager need to support a changelog topic with a topic set to None.

* Added support for custom changelog topic configuration (instead of relying on the "data" topic configuration)

When creating the changelog topic we use a configuration similar to the data topic configuration (replication factor, number of partitions, ...). As the data topic is in-existant for stateful sources we implement a way to create a changelog topic with a passed configuration.