Correct _get_items documentation
The helper method `_get_items` is more in line with
a native rocksdict behavior where the lower band is
inclusive and the upper exclusive.

The inverted characteristics known from an upper
scope's `get_windows` method are guaranteed by key
deserialization and condition:
`start_from_ms < start <= start_to_ms`.

In the future, we should refactor the code to remove
this condition and rely on `_get_items`
inclusivity/exclusivity characteristics.