Return empty list if the start is higher than the end
It is a shortcut to empty results if the given
start timestamp is higher than the end timestamp.

There are currently use cases in the code that may
lead to this and there is no easy way to avoid it.

The shortcut is required to avoid unexpected behaviors
with negative integers converted to int64 encoded bytes.