Feature: Collection Aggregations (#688)
* Remove unnecessary value from docstring

* Rename encode_window_prefix to a more generic append_integer

* Invert condition in expire_windows

* Move max window timestamp logic from state to window

* Reorder value and timestamp_ms arguments for consistency

* Extract _get_items from get_windows method

* Clarify typing in window definitions

* Reorder arguments in FixedTimeWindow

* Create collect aggregation function and pass aggregate_collection arg

* Correct typing in serialization.py

* Rename PREFIX_SEPARATOR to SEPARATOR

* Remove redundant cast to list

* Feature: Collection Aggregations

* Add latest deleted value cache

* Set last deleted value timestamp

* Minor fixes in sliding windows tests

* Make sliding windows tests specific for reduce aggregation function

* Deserialize and re-serialize key when deleting value

This is dictated by the legacy structure of the code.
Method `delete` requires prefix and remainder of a key to be passed
separately and refactoring this is a larger effort.

* Correct _get_items documentation

The helper method `_get_items` is more in line with
a native rocksdict behavior where the lower band is
inclusive and the upper exclusive.

The inverted characteristics known from an upper
scope's `get_windows` method are guaranteed by key
deserialization and condition:
`start_from_ms < start <= start_to_ms`.

In the future, we should refactor the code to remove
this condition and rely on `_get_items`
inclusivity/exclusivity characteristics.

* Return empty list if the start is higher than the end

It is a shortcut to empty results if the given
start timestamp is higher than the end timestamp.

There are currently use cases in the code that may
lead to this and there is no easy way to avoid it.

The shortcut is required to avoid unexpected behaviors
with negative integers converted to int64 encoded bytes.

* Add collection aggregation sliding windows test

* Helper variable in FixedTimeWindow.process_window

* Fix mypy

* Add test_collect_value

* Add test_delete_windows_with_values

* Add test_expire_windows_with_collect

* Add test_tumblingwindow_collect

* Add test_hoppingwindow_collect

* Fix docstrings and comments

* Add entry in windowing documentation

* Ensure that values coming at the same time are collected in correct order

* Use direct FixedTimeWindow type hint

* Add comment explaining why window values are None

* Rename collect_value to add_to_collection